requirements.txt
# Core dependencies
pyserial>=3.5
numpy>=1.24.0

# Voice processing
openwakeword>=0.5.0
whisper-openai>=20230314
piper-tts>=1.2.0
pyaudio>=0.2.13

# Machine Learning / AI
torch>=2.0.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0

# HTTP requests
requests>=2.31.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional GPU acceleration (uncomment if using CUDA)
# torch>=2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
